pyproject.toml
[tool.poetry]
name = "autothumbnailgenerator"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.84"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
